+++ editor/src/core/shared/math-utils.ts
@@ -1024,7 +1024,3 @@ export function segmentIntersection(
 export function mod(n: number, m: number) {
   return ((n % m) + m) % m
 }
-
-export function easeOutCubic(x: number): number {
-  return 1 - Math.pow(1 - x, 3)
-}
